Fix Create new App in Portal
