+++ src/appcenter/commands/appCenterPortalMenu.ts
@@ -1,5 +1,5 @@
 import * as vscode from "vscode";
-import { AppCenteAppType } from "../../constants";
+import { AppCenteAppType, CommandNames } from "../../constants";
 import { CommandParams, CurrentApp, QuickPickAppItem } from "../../helpers/interfaces";
 import { MenuHelper } from "../../helpers/menuHelper";
 import { Utils } from "../../helpers/utils";
@@ -12,11 +12,9 @@ export default class AppCenterPortalMenu extends ReactNativeAppCommand {
     private appName: string;
     private ownerName: string;
     private isOrg: boolean;
-    private _params: CommandParams;
 
     constructor(params: CommandParams) {
         super(params);
-        this._params = params;
     }
 
     public async run(): Promise<void> {
@@ -31,34 +29,39 @@ export default class AppCenterPortalMenu extends ReactNativeAppCommand {
     }
 
     protected async handleShowCurrentAppQuickPickSelection(selected: QuickPickAppItem, rnApps: models.AppResponse[]) {
-        let selectedApp: models.AppResponse;
-
-        const selectedApps: models.AppResponse[] = rnApps.filter(app => app.name === selected.target && app.owner.type === selected.description);
-
-        // If this is not current app then we can assign current app, otherwise we will use GetCurrentApp method
-        if (selected.target !== this.currentAppMenuTarget) {
-            if (!selectedApps || selectedApps.length !== 1) {
-                return;
-            }
-            selectedApp = selectedApps[0];
-        }
-
-        if (selectedApp) {
-            this.isOrg = selectedApp.owner.type.toLowerCase() === AppCenteAppType.Org.toLowerCase();
-            this.appName = selectedApp.name;
-            this.ownerName = selectedApp.owner.name;
+        if (selected.target === CommandNames.CreateApp.CommandName) {
+            return this.showCreateAppOptions();
         } else {
-            const currentApp: CurrentApp | null = await this.getCurrentApp();
-            if (currentApp) {
-                this.isOrg = currentApp.type.toLowerCase() === AppCenteAppType.Org.toLowerCase();
-                this.appName = currentApp.appName;
-                this.ownerName = currentApp.ownerName;
-            } else {
-                this.logger.error("Current app is undefiend");
-                throw new Error("Current app is undefiend");
+
+            let selectedApp: models.AppResponse;
+
+            const selectedApps: models.AppResponse[] = rnApps.filter(app => app.name === selected.target && app.owner.type === selected.description);
+
+            // If this is not current app then we can assign current app, otherwise we will use GetCurrentApp method
+            if (selected.target !== this.currentAppMenuTarget) {
+                if (!selectedApps || selectedApps.length !== 1) {
+                    return;
+                }
+                selectedApp = selectedApps[0];
             }
+
+            if (selectedApp) {
+                this.isOrg = selectedApp.owner.type.toLowerCase() === AppCenteAppType.Org.toLowerCase();
+                this.appName = selectedApp.name;
+                this.ownerName = selectedApp.owner.name;
+            } else {
+                const currentApp: CurrentApp | null = await this.getCurrentApp();
+                if (currentApp) {
+                    this.isOrg = currentApp.type.toLowerCase() === AppCenteAppType.Org.toLowerCase();
+                    this.appName = currentApp.appName;
+                    this.ownerName = currentApp.ownerName;
+                } else {
+                    this.logger.error("Current app is undefiend");
+                    throw new Error("Current app is undefiend");
+                }
+            }
+            this.showAppCenterPortalMenuQuickPick(MenuHelper.getAppCenterPortalMenuItems(Utils.isReactNativeProject(this.logger, this.rootPath, false)));
         }
-        this.showAppCenterPortalMenuQuickPick(MenuHelper.getAppCenterPortalMenuItems(Utils.isReactNativeProject(this.logger, this.rootPath, false)));
     }
 
     private async showAppCenterPortalMenuQuickPick(appCenterMenuOptions: CustomQuickPickItem[]): Promise<void> {
